+++ app.php
@@ -6,7 +6,7 @@
 
 use Slim\Factory\AppFactory;
 
-define('KUUSAMO_VERSION', '1.8.2');
+define('KUUSAMO_VERSION', '1.9.0');
 
 function kuusamo()
 {