+++ src/Migration/Version20201229173109.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kuusamo\Vle\Migration;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+final class Version20201229173109 extends AbstractMigration
+{
+    public function getDescription() : string
+    {
+        return '';
+    }
+
+    public function up(Schema $schema) : void
+    {
+        $this->addSql('CREATE TABLE vle_settings (`name` VARCHAR(128) NOT NULL, value LONGTEXT NOT NULL, PRIMARY KEY(`name`)) DEFAULT CHARACTER SET utf8 COLLATE `utf8_unicode_ci` ENGINE = InnoDB');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->addSql('DROP TABLE vle_settings');
+    }
+}